+++ frmtr-gradle-plugin/src/main/java/dev/lanwen/frmtr/gradle/FrmtrExtension.java
@@ -3,8 +3,11 @@ package dev.lanwen.frmtr.gradle;
 import javax.inject.Inject;
 import org.gradle.api.Action;
 import org.gradle.api.model.ObjectFactory;
+import org.gradle.api.provider.Property;
 
 public abstract class FrmtrExtension {
+
+    private final Property<Boolean> enabled;
 
     private final FrmtrJavaExtension java;
 
@@ -12,8 +15,13 @@ public abstract class FrmtrExtension {
 
     @Inject
     public FrmtrExtension(ObjectFactory objects) {
+        this.enabled = objects.property(Boolean.class).convention(true);
         this.java = objects.newInstance(FrmtrJavaExtension.class);
         this.check = objects.newInstance(FrmtrCheckExtension.class);
+    }
+
+    public Property<Boolean> getEnabled() {
+        return enabled;
     }
 
     public FrmtrJavaExtension getJava() {